+++ deploy/src/bin/neutron_initialization.rs
@@ -4,7 +4,6 @@ use cosmwasm_std::Binary;
 use serde::Deserialize;
 use sp1_sdk::{HashableKey, SP1VerifyingKey};
 use types::{
-    gaia_config::GaiaStrategyConfig,
     labels::{
         ICA_TRANSFER_LABEL, LEND_AND_PROVIDE_LIQUIDITY_LABEL, MARS_WITHDRAW_LABEL,
         PHASE_SHIFT_LABEL, REGISTER_OBLIGATION_LABEL, SETTLE_OBLIGATION_LABEL,
@@ -29,7 +28,6 @@ use valence_library_utils::LibraryAccountType;
 #[derive(Deserialize, Debug)]
 struct Parameters {
     general: General,
-    ica: Ica,
     coprocessor_app: CoprocessorApp,
 }
 
@@ -37,11 +35,6 @@ struct Parameters {
 struct General {
     owner: String,
     strategist: String,
-}
-
-#[derive(Deserialize, Debug)]
-struct Ica {
-    deposit_token_on_hub_denom: String,
 }
 
 #[derive(Deserialize, Debug)]
@@ -559,52 +552,6 @@ async fn main() -> Result<(), Box<dyn Error>> {
     println!("ZK Authorization created successfully");
 
     // TODO: Transfer ownership of authorization contract to owner
-
-    // Last thing we will do is register the ICA on the valence ICA
-    let register_ica_msg = valence_account_utils::ica::ExecuteMsg::RegisterIca {};
-    neutron_client
-        .execute_wasm(
-            &ntrn_strategy_config.accounts.ica,
-            register_ica_msg,
-            vec![cosmrs::Coin::new(1_000_000u128, "untrn").unwrap()],
-            None,
-        )
-        .await?;
-
-    println!("Registering ICA");
-
-    // Let's wait enough time for the transaction to succeed and the ICA to be registered
-    tokio::time::sleep(std::time::Duration::from_secs(30)).await;
-
-    // Let's query now to get the ICA address
-    let query_ica = valence_account_utils::ica::QueryMsg::IcaState {};
-    let ica_state: valence_account_utils::ica::IcaState = neutron_client
-        .query_contract_state(&ntrn_strategy_config.accounts.ica, query_ica)
-        .await?;
-
-    let ica_address = match ica_state {
-        valence_account_utils::ica::IcaState::Created(ica_information) => ica_information.address,
-        _ => {
-            panic!("ICA creation failed!");
-        }
-    };
-
-    let gaia_cfg = GaiaStrategyConfig {
-        grpc_url: "grpc_url".to_string(),
-        grpc_port: "grpc_port".to_string(),
-        chain_id: "chain_id".to_string(),
-        chain_denom: "uatom".to_string(),
-        deposit_denom: ntrn_params.ica.deposit_token_on_hub_denom,
-        ica_address,
-    };
-
-    // Write the Gaia strategy config to a file
-    let gaia_cfg_path = current_dir.join("deploy/src/gaia_strategy_config.toml");
-    fs::write(
-        gaia_cfg_path,
-        toml::to_string(&gaia_cfg).expect("Failed to serialize Gaia strategy config"),
-    )
-    .expect("Failed to write Gaia strategy config to file");
 
     Ok(())
 }
