+++ deploy/src/ethereum.toml
@@ -1,20 +1,20 @@
 [general]
 rpc_url = "http://localhost:8545"
-owner   = "0x54a37ac81263C482D6BE56F5Bd796e06e9Afa344" # Neutron multisig address
-valence_owner = "0x54a37ac81263C482D6BE56F5Bd796e06e9Afa344" # TBD Owner of the verification gateway
+owner   = "0xd9A23b58e684B985F661Ce7005AA8E10630150c1" # Neutron multisig address
+valence_owner = "0xd9A23b58e684B985F661Ce7005AA8E10630150c1" # TBD Owner of the verification gateway
 coprocessor_root = "0x0000000000000000000000000000000000000000000000000000000000000000" # Bytes32 hash of coprocessor root
 
 [vault]
-deposit_token            = "0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599"
-strategist               = "0x54a37ac81263C482D6BE56F5Bd796e06e9Afa344" # Address of the strategist
-platform_fee_account     = "0x54a37ac81263C482D6BE56F5Bd796e06e9Afa344" # Receiver of the platform fees portion
-strategist_fee_account   = "0x54a37ac81263C482D6BE56F5Bd796e06e9Afa344" # Receiver of the strategist fees portion
-strategist_fee_ratio_bps = 500                                          # Strategist fee is 5%, rest goes to the platform (95%)
+deposit_token            = "0x8236a87084f8B84306f72007F36F2618A5634494"
+strategist               = "0x510c4a1d637ff374399826f421003b775dc3e8dc" # Address of the strategist
+platform_fee_account     = "0x510c4a1d637ff374399826f421003b775dc3e8dc" # Receiver of the platform fees portion
+strategist_fee_account   = "0x510c4a1d637ff374399826f421003b775dc3e8dc" # Receiver of the strategist fees portion
+strategist_fee_ratio_bps = 5000                                         # Strategist fee is 50%, rest goes to the platform (50%)
 deposit_cap              = "10000000000000"                             # e.g. 10K BTC
 deposit_fee_bps          = 100                                          # 1% deposit fee
 withdraw_rate_bps        = 100                                          # 1% withdrawal rate
-starting_rate            = "10000000"                                   # 1e8 - WBTC Precision
-max_rate_update_delay    = 172800                                       # 48 hours in seconds - Maximum delay for rate updates before vault gets paused
+starting_rate            = "100000000"                                  # 1e8 - WBTC Precision
+max_rate_update_delay    = 345600                                       # 4 days in seconds - Maximum delay for rate updates before vault gets paused
 rate_scaling_factor      = 1000000000000                                # scaling factor for rate updates
 
 [eureka_transfer]
